[webdriver] Convert inline & iframe to fixtures
